--- a/src/test/regress/optfunctional/.gitignore
+++ b/src/test/regress/optfunctional/.gitignore
@@ -1,0 +1,7 @@
+pg_regress
+results
+yml
+current_good_schedule
+optimizer_status.out
+regression.diffs
+regression.out