… src/test/regress/optfunctional/data/world.sql
… src/test/regress/optfunctional/expected/create_table_default_distribution_on.out
… src/test/regress/optfunctional/expected/cte_functest_inlining_disabled.out
… src/test/regress/optfunctional/expected/cte_functest_inlining_enabled.out
… src/test/regress/optfunctional/expected/cte_queries_inlining_disabled.out
… src/test/regress/optfunctional/expected/cte_queries_inlining_enabled.out
… src/test/regress/optfunctional/expected/lastj.out
… src/test/regress/optfunctional/expected/lastj_hash.out
… src/test/regress/optfunctional/expected/lineitem.out
… src/test/regress/optfunctional/expected/partitionindexes.out
… src/test/regress/optfunctional/expected/staticselection.out
… src/test/regress/optfunctional/init_file
… src/test/regress/optfunctional/input/create_table_as_11_alter_table.source
… src/test/regress/optfunctional/input/cte_queries_inlining_disabled.source
… src/test/regress/optfunctional/input/cte_queries_inlining_enabled.source
… src/test/regress/optfunctional/input/indexapply.source
… src/test/regress/optfunctional/mapred/dummy.source
… src/test/regress/optfunctional/optfunctional_schedule
… src/test/regress/optfunctional/output/composite_keys_gpdb_1.source
… src/test/regress/optfunctional/output/composite_keys_gpdb_2.source
… src/test/regress/optfunctional/output/composite_keys_gpdb_3.source
… src/test/regress/optfunctional/output/create_table_as_11_alter_table.source
… src/test/regress/optfunctional/output/indexapply.source
… src/test/regress/optfunctional/sql/analyze_rootpartition.sql
… src/test/regress/optfunctional/sql/create_table_default_distribution_off.sql
… src/test/regress/optfunctional/sql/create_table_default_distribution_on.sql
… src/test/regress/optfunctional/sql/cte_functest_inlining_disabled.sql
… src/test/regress/optfunctional/sql/cte_functest_inlining_enabled.sql
… src/test/regress/optfunctional/sql/lastj.sql
… src/test/regress/optfunctional/sql/lastj_hash.sql