Includes tests:
  composite_keys_gpdb_1
  composite_keys_gpdb_3
  composite_keys_gpdb_2
  partitionindexes
  staticselection
  cte_functest
  cte_queries
  create_table_default_distribution
  lastj lastj_hash
  indexapply

NOTE :
1. Added a new target installcheck-optfunctional
2. analyze_rootpartition ... fails on master
3. optimizer functional tests are in src/test/regress/optfunctional directory.

```
[/workspace/tinc/gpdb/src/test/regress] [00:26:44] ip-172-31-5-70.us-west-1.compute.internal (0)
$ PGOPTIONS="-c optimizer=on" make installcheck-optfunctional
make -C ../../../src/port all
make[1]: Entering directory `/workspace/tinc/gpdb/src/port'
make[1]: Nothing to be done for `all'.
make[1]: Leaving directory `/workspace/tinc/gpdb/src/port'
make -C ../../../contrib/spi refint.so autoinc.so
make[1]: Entering directory `/workspace/tinc/gpdb/contrib/spi'
make[1]: `refint.so' is up to date.
make[1]: `autoinc.so' is up to date.
make[1]: Leaving directory `/workspace/tinc/gpdb/contrib/spi'
rm -rf ./testtablespace
mkdir ./testtablespace
./checkinc.py
Greenplum INCLUDEDIR:    /workspace/tinc/gpdb/build/include
Greenplum PKGINCLUDEDIR: /workspace/tinc/gpdb/build/include/postgresql
Checking includes...
Include files are ok
================================
Installcheck-optfunctional starting ...
================================
./pg_regress --inputdir=./optfunctional --outputdir=./optfunctional --multibyte=   --init-file=./init_file --psqldir=/workspace/tinc/gpdb/build/bin --sched
ule=./optfunctional/optfunctional_schedule --srcdir=/workspace/tinc/gpdb/src/test/regress/optfunctional
(using postmaster on Unix socket, port 9876)
l============== dropping database "regression"         ==============
DROP DATABASE
============== creating database "regression"         ==============
CREATE DATABASE
ALTER DATABASE
============== checking optimizer status              ==============
Optimizer enabled. Using optimizer answer files whenever possible
============== running regression test queries        ==============
parallel group (3 tests):  composite_keys_gpdb_2 composite_keys_gpdb_1 composite_keys_gpdb_3
     composite_keys_gpdb_1 ... ok (0.14 sec)  (diff:0.05 sec)
     composite_keys_gpdb_2 ... ok (0.13 sec)  (diff:0.05 sec)
     composite_keys_gpdb_3 ... ok (0.14 sec)  (diff:0.05 sec)
test partitionindexes     ... ok (68.48 sec)  (diff:0.20 sec)
test staticselection      ... ok (1.15 sec)  (diff:0.09 sec)
parallel group (2 tests):  cte_functest_inlining_enabled cte_queries_inlining_enabled
     cte_functest_inlining_enabled ... ok (21.34 sec)  (diff:0.25 sec)
     cte_queries_inlining_enabled ... ok (21.73 sec)  (diff:0.20 sec)
parallel group (2 tests):  cte_queries_inlining_disabled cte_functest_inlining_disabled
     cte_queries_inlining_disabled ... ok (21.52 sec)  (diff:0.20 sec)
     cte_functest_inlining_disabled ... ok (25.77 sec)  (diff:0.25 sec)
parallel group (7 tests):  create_table_default_distribution_off create_table_as_11_alter_table create_table_default_distribution_on lastj lastj_hash index
apply lineitem
     lastj                ... ok (2.31 sec)  (diff:0.11 sec)
     lastj_hash           ... ok (2.44 sec)  (diff:0.10 sec)
     lineitem             ... ok (16.05 sec)  (diff:1.58 sec)
     create_table_default_distribution_on ... ok (1.73 sec)  (diff:0.07 sec)
     create_table_default_distribution_off ... ok (0.33 sec)  (diff:0.05 sec)
     indexapply           ... ok (4.79 sec)  (diff:0.08 sec)
     create_table_as_11_alter_table ... ok (1.29 sec)  (diff:0.06 sec)
test analyze_rootpartition ... FAILED (7.19 sec)  (diff:0.14 sec)

=======================
 1 of 17 tests failed.
=======================
```