@danielgustafsson 2016-05-24
The lineitem test adds ~8.5Mb to the repo size and these increases will over time slow down git clone and other operations in the CI processes, is it possible to shrink this a bit before pushing (once pushed it will be in the repo forever)? Our current `src/test/regress` is ~69Mb so it's a sizeable increase.