@hardikar 2016-05-24
@danielgustafsson In fact that is shared with bugbuster. Even the world.sql is in fact shared with bugbuster. I will try to move them to a common folder (possibly the src/test/regress/data) from which both these test suites can access it.