@hardikar 2016-05-25
@danielgustafsson Have a look also at https://github.com/greenplum-db/gpdb/pull/785. I moved it to the csv files to the parent folder. That should save us 11 MB. ```lineitem.out``` and ```world.sql``` are 1.6M and 196K respectively - working to see if those can be easily ported/reduced in size.