@danielgustafsson 2016-05-25
Nice catch! Re-using the data files like that will not add to the size of the repo for the input. If you can find a way to reduce the size of the `lineitem.out` file as well then it will be even better. Great work!