@oarap 2016-05-31
LGTM after the comments addressed in the PR.