@hardikar 2016-05-31
Addressed PR comments from @oarap in bc9413e23f89050365d822d5d0a607193f625190.