@hardikar 2016-06-01
@danielgustafsson Spoke with the QO team, looks like we need the lineitem.out file. Thanks for review.